Начало кружки

+++ api_rest.py
@@ -5,14 +5,12 @@ from pprint import pprint
 from loguru import logger
 import requests
 
-from config import ready_path, main_path, sticker_path
+from config import ready_path, main_path, sticker_path, ready_path_kruzhka
 from utils.created_pdf import one_pdf
 
 headers = {'Content-Type': 'application/json'}
-# domain = 'http://127.0.0.1:8000/api_rest'
-
-
-domain = 'https://mycego.online/api_rest'
+domain = 'http://127.0.0.1:8000/api_rest'
+# domain = 'https://mycego.online/api_rest'
 
 
 def get_products(categories: list):
@@ -25,36 +23,38 @@ def get_products(categories: list):
         logger.error(f'Ошибка в запросе по api {ex}')
 
 
-def get_info_publish_folder(public_url, files):
+def get_info_publish_folder(public_url):
     result_data = []
     res = requests.get(
         f'https://cloud-api.yandex.net/v1/disk/public/resources?public_key={public_url}&fields=_embedded&limit=1000')
     if res.status_code == 200:
         data = res.json().get('_embedded', {}).get('items', [])
         for i in data:
-            file_name = i.get('name', None)
-            if file_name in files:
-                try:
-                    result_data.append({
-                        'name': i.get('name'),
-                        'file': i.get('file')
-                    })
-                except:
-                    pass
+            try:
+                result_data.append({
+                    'name': i.get('name'),
+                    'file': i.get('file')
+                })
+            except:
+                pass
         return result_data
 
 
 def create_download_data(item):
-    download_files = []
     if len(item.get('images')) != item['quantity'] and not item['the_same']:
         logger.error(f'Не совпадает количество\n{item}')
     else:
-        download_files.extend(item['images'])
-        download_files.extend(item['sticker'])
-
-        url_data = get_info_publish_folder(item['directory_url'], download_files)
+        result = []
+        url_data = get_info_publish_folder(item['directory_url'])
+        print(url_data)
         if url_data:
-            item['url_data'] = url_data
+            for item in url_data:
+                item_name: str = item['name']
+                if item_name.split('.')[0].isdigit() or 'принт' in item_name.lower() or (
+                        item_name.endswith('.pdf') and 'изобра' not in item_name):
+                    result.append(item)
+
+            item['url_data'] = result
             return item
 
 
@@ -91,12 +91,12 @@ def main_download_site(categories, dir_path):
     logger.debug(f'Артикулов в ответе с сайта:{len(data)}')
     data = [item for item in data if item['art'] not in art_list]
     logger.success(f'Артикулов для загрузки:{len(data)}')
-    for item in data:
+    for item in data[:2]:
         download_data = create_download_data(item)
         if download_data:
             result_dict_arts.append(download_data)
-    # with open('json.json', 'w') as f:
-    #     json.dump(result_dict_arts, f, indent=4, ensure_ascii=False)
+    with open(f'json {categories[0]}.json', 'w') as f:
+        json.dump(result_dict_arts, f, indent=4, ensure_ascii=False)
     #
     # with open('json.json', 'r') as f:
     #     data = json.load(f)
@@ -109,45 +109,54 @@ def main_download_site(categories, dir_path):
         try:
             if int(art.split('-')[-1]) != count:
                 logger.error(f'Не совпадает кол-во {art}')
+                continue
         except Exception as ex:
             logger.error(f'Не понятный {art} {ex}')
-        os.makedirs(folder, exist_ok=True)
-        for i in item['url_data']:
-            destination_path = os.path.join(folder, i['name'])
-            download_file(destination_path, i['file'])
+            continue
+        if 'Кружки' in categories:
+            for i in item['url_data']:
+                exep = i['name'].split('.')[-1]
+                destination_path = os.path.join(dir_path, f'{art}.{exep}')
+                download_file(destination_path, i['file'])
+        else:
+            os.makedirs(folder, exist_ok=True)
+            for i in item['url_data']:
+                destination_path = os.path.join(folder, i['name'])
+                download_file(destination_path, i['file'])
 
-        if item['the_same']:
-            try:
-                image_path = os.path.join(folder, '1.png')
-                if os.path.exists(image_path):
-                    copy_image(image_path, count)
-                else:
-                    image_path = os.path.join(folder, '1.jpg')
+            if item['the_same']:
+                try:
+                    image_path = os.path.join(folder, '1.png')
                     if os.path.exists(image_path):
                         copy_image(image_path, count)
                     else:
-                        raise ValueError(f'Нет файла для копирования артикул: {item["art"]}')
+                        image_path = os.path.join(folder, '1.jpg')
+                        if os.path.exists(image_path):
+                            copy_image(image_path, count)
+                        else:
+                            raise ValueError(f'Нет файла для копирования артикул: {item["art"]}')
+                except Exception as ex:
+                    logger.error(ex)
+
+            try:
+                for file in os.listdir(folder):
+                    if file.endswith('.pdf'):
+                        try:
+                            shutil.move(os.path.join(folder, file), sticker_path)
+                        except Exception as ex:
+                            logger.error(ex)
+                            os.remove(os.path.join(folder, file))
+                filename = f'{dir_path}\\{art}.pdf'
+                logger.debug(filename)
+                if not os.path.exists(filename) and os.path.isdir(folder):
+                    one_pdf(folder_path=folder, art=art)
+                else:
+                    logger.error(f'файл существует {filename}')
+                logger.success(f'{index}/{count_task} - {item["art"]}')
             except Exception as ex:
                 logger.error(ex)
-
-        try:
-            for file in os.listdir(folder):
-                if file.endswith('.pdf'):
-                    try:
-                        shutil.move(os.path.join(folder, file), sticker_path)
-                    except Exception as ex:
-                        logger.error(ex)
-                        os.remove(os.path.join(folder, file))
-            filename = f'{dir_path}\\{art}.pdf'
-            logger.debug(filename)
-            if not os.path.exists(filename) and os.path.isdir(folder):
-                one_pdf(folder_path=folder, art=art)
-            else:
-                logger.error(f'файл существует {filename}')
-            logger.success(f'{index}/{count_task} - {item["art"]}')
-        except Exception as ex:
-            logger.error(ex)
+            break
 
 
 if __name__ == '__main__':
-    main_download_site()
+    main_download_site(categories=['Кружки'], dir_path=ready_path_kruzhka)
